Move verifyProof logic to own file
